Changed the validation error messages, moved cain to components, changed the speech bubble to be center aligned
